+++ example/lib/main.dart
@@ -1,9 +1,30 @@
 import 'package:flutter/material.dart';
 
 import 'package:qz/qz.dart';
+import 'package:qz_example/resources.dart';
 
 void main() {
   runApp(const MyApp());
+}
+
+class PrinterWrapper {
+  final String title;
+  final String? name;
+
+  const PrinterWrapper({
+    required this.title,
+    this.name,
+  });
+
+  @override
+  bool operator ==(Object other) =>
+      identical(this, other) ||
+      other is PrinterWrapper &&
+          runtimeType == other.runtimeType &&
+          name == other.name;
+
+  @override
+  int get hashCode => name?.hashCode ?? 0;
 }
 
 class MyApp extends StatefulWidget {
@@ -23,14 +44,14 @@ class _MyAppState extends State<MyApp> {
     ''',
   );
 
-  String? printer;
-  String pdfBase64 = '''
-^XA
-^FO50,60^A0,40^FDPrinting ZPL^FS
-^FO60,120^BY3^BCN,60,,,,A^FD1234ABC^FS
-^FO25,25^GB380,200,2^FS
-^XZ
+  static const defaultPrinter = PrinterWrapper(title: 'Use default printer');
+  PrinterWrapper? printer = defaultPrinter;
+  String? pdfBase64Input;
+  String zplSample = '''
+^XA^FO17,16^GB379,371,8^FS^FT65,255^A0N,135,134^FDTEST^FS^XZ
   ''';
+
+  String? errorText;
 
   @override
   void initState() {
@@ -40,63 +61,142 @@ class _MyAppState extends State<MyApp> {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      home: Scaffold(
-        appBar: AppBar(
-          title: Text('Running QZ Tray ${_qzWebPlugin.getQzVersion()}'),
-        ),
-        body: Center(
-          child: Column(
-            mainAxisAlignment: MainAxisAlignment.start,
-            mainAxisSize: MainAxisSize.min,
-            children: [
-              FutureBuilder(
-                  future: _qzWebPlugin.getAllPrinters(),
-                  builder: (context, snapshot) {
-                    if (snapshot.hasData) {
-                      return DropdownButtonFormField(
-                          value: printer,
-                          items: snapshot.data
-                              ?.map((element) => DropdownMenuItem(
-                                    value: element,
-                                    child: Text(element),
-                                  ))
-                              .toList(),
-                          onChanged: (value) {
-                            setState(() {
-                              printer = value;
-                            });
-                          });
-                    }
-                    return const SizedBox();
-                  }),
-              TextField(
-                decoration: const InputDecoration(
-                  hintText: 'Base64 string pdf',
-                ),
-                onChanged: (value) {
-                  pdfBase64 = value;
-                },
-              ),
-              const SizedBox(height: 16),
-              SizedBox(
-                width: 100,
-                child: ElevatedButton(
-                    onPressed: () {
-                      _qzWebPlugin.print(
-                          printerName: printer, base64: pdfBase64);
+      home: SafeArea(
+        child: Scaffold(
+          appBar: AppBar(
+            title: Text('Running QZ Tray ${_qzWebPlugin.version}'),
+          ),
+          body: Padding(
+            padding: const EdgeInsets.all(16.0),
+            child: Center(
+              child: Column(
+                mainAxisAlignment: MainAxisAlignment.start,
+                mainAxisSize: MainAxisSize.min,
+                children: [
+                  FutureBuilder(
+                      future: _qzWebPlugin
+                          .getAllPrinters()
+                          .timeout(const Duration(seconds: 5)),
+                      builder: (context, snapshot) {
+                        if (snapshot.hasData) {
+                          final List<PrinterWrapper> printers = [
+                            defaultPrinter,
+                            ...snapshot.data
+                                    ?.map((item) =>
+                                        PrinterWrapper(title: item, name: item))
+                                    .toList() ??
+                                [],
+                          ];
+                          return DropdownButtonFormField<PrinterWrapper>(
+                              value: printer,
+                              items: printers
+                                  .map((printer) => DropdownMenuItem(
+                                        value: printer,
+                                        child: Text(printer.title),
+                                      ))
+                                  .toList(),
+                              onChanged: (value) {
+                                setState(() {
+                                  printer = value;
+                                });
+                              });
+                        }
+                        if (snapshot.connectionState ==
+                            ConnectionState.waiting) {
+                          return const LinearProgressIndicator();
+                        }
+                        return Column(
+                          mainAxisSize: MainAxisSize.min,
+                          children: [
+                            if (snapshot.hasError) ...[
+                              Text(
+                                '${snapshot.error?.toString()}\n Check you are using the correct "certificateString" and "signatureString" when building QZ plugin.',
+                                textAlign: TextAlign.center,
+                                style: Theme.of(context)
+                                    .textTheme
+                                    .titleMedium
+                                    ?.copyWith(
+                                        color: Theme.of(context)
+                                            .colorScheme
+                                            .error),
+                              ),
+                              const SizedBox(height: 8),
+                            ],
+                            ElevatedButton(
+                              onPressed: () {
+                                setState(() {});
+                              },
+                              child: const Text('Reload printers'),
+                            ),
+                          ],
+                        );
+                      }),
+                  TextField(
+                    decoration: InputDecoration(
+                      hintText: 'Base64 string pdf',
+                      errorText: errorText,
+                    ),
+                    onChanged: (value) {
+                      pdfBase64Input = value;
                     },
-                    child: const Text('Print')),
+                  ),
+                  const SizedBox(height: 16),
+                  ElevatedButton(
+                      onPressed: () {
+                        if (pdfBase64Input?.isEmpty ?? true) {
+                          errorText = 'Base 64 PDF input can not be empty';
+                          setState(() {});
+                          return;
+                        } else {
+                          errorText = null;
+                          setState(() {});
+                        }
+                        _qzWebPlugin.printPDF(
+                            printerName: printer?.name, base64: pdfBase64Input);
+                      },
+                      child: const Text('Print PDF From Base64 Input')),
+                  const SizedBox(height: 16),
+                  ElevatedButton(
+                      onPressed: () {
+                        _qzWebPlugin.printPDF(
+                            printerName: printer?.name,
+                            base64: pdfBase64Sample);
+                      },
+                      child: const Text('Print PDF Bas64 Sample')),
+                  const SizedBox(height: 16),
+                  ElevatedButton(
+                      onPressed: () {
+                        _qzWebPlugin.printPDF(
+                            printerName: printer?.name,
+                            uri: Uri.tryParse('assets/ticket.pdf'));
+                      },
+                      child: const Text('Print PDF File Sample')),
+                  const SizedBox(height: 16),
+                  ElevatedButton(
+                      onPressed: () {
+                        _qzWebPlugin.printRaw(
+                            printerName: printer?.name,
+                            base64: pdfBase64Sample);
+                      },
+                      child: const Text('Print PDF Base64 Sample Raw')),
+                  const SizedBox(height: 16),
+                  ElevatedButton(
+                      onPressed: () {
+                        _qzWebPlugin.printRaw(
+                            printerName: printer?.name,
+                            uri: Uri.tryParse('assets/ticket.pdf'));
+                      },
+                      child: const Text('Print PDF File Sample Raw')),
+                  const SizedBox(height: 16),
+                  ElevatedButton(
+                      onPressed: () {
+                        _qzWebPlugin.printZPL(
+                            printerName: printer?.name, zpl: zplSample);
+                      },
+                      child: const Text('Print ZPL Sample')),
+                ],
               ),
-              const SizedBox(height: 16),
-              SizedBox(
-                width: 100,
-                child: ElevatedButton(
-                    onPressed: () {
-                      _qzWebPlugin.printZpl(zplString: pdfBase64);
-                    },
-                    child: const Text('Print ZPL')),
-              ),
-            ],
+            ),
           ),
         ),
       ),