+++ lib/qz.dart
@@ -3,17 +3,10 @@ import 'package:qz/qz_stub.dart'
 
 class Qz extends QzWeb {
   Qz(
-      {bool backendMode = false,
-      String algorithm = 'SHA512',
-      String? certificateUrl,
-      String? signatureUrl,
-      String? certificateString,
-      String? signatureString})
-      : super(
-            backendMode: backendMode,
-            algorithm: algorithm,
-            certificateUrl: certificateUrl,
-            signatureUrl: signatureUrl,
-            certificateString: certificateString,
-            signatureString: signatureString);
+      {super.backendMode,
+      super.algorithm,
+      super.certificateUrl,
+      super.signatureUrl,
+      super.certificateString,
+      super.signatureString});
 }